# Implement codegen from regex DFA.