--- a/dialects/regex/pass/dfa.cpp
+++ b/dialects/regex/pass/dfa.cpp
@@ -25,7 +25,7 @@ void DFANode::dump() const {
         if (c == DFA::SpecialTransitons::ANY)
             return ".";
         else
-            return {1, static_cast<char>(c)};
+            return {static_cast<char>(c)};
     };
 
     if (this->is_accepting()) std::cout << "  \"" << this << "\" [shape=doublecircle];\n";