--- a/dialects/regex/pass/dfa.cpp
+++ b/dialects/regex/pass/dfa.cpp
@@ -2,23 +2,15 @@
 
 #include <iostream>
 
+#include "thorin/util/util.h"
+
 #include "dialects/regex/pass/automaton.h"
 
 using namespace thorin::automaton;
 
-void DFANode::add_transition(const DFANode* to, std::uint16_t c) {
-    if (auto it = transitions_.find(c); it != transitions_.end())
-        it->second = to;
-    else
-        transitions_.emplace(c, to);
-}
+void DFANode::add_transition(const DFANode* to, std::uint16_t c) { transitions_[c] = to; }
 
-const DFANode* DFANode::get_transition(std::uint16_t c) const {
-    if (auto it = transitions_.find(c); it != transitions_.end())
-        return it->second;
-    else
-        return {};
-}
+const DFANode* DFANode::get_transition(std::uint16_t c) const { return lookup(transitions_, c); }
 
 void DFANode::dump() const {
     auto print_char = [](std::uint16_t c) -> std::string {