--- a/dialects/regex/pass/dfa.h
+++ b/dialects/regex/pass/dfa.h
@@ -16,19 +16,13 @@ public:
     const DFANode* get_transition(std::uint16_t c) const;
 
     // F: void(const DFANode*)
-    template <class F>
-    void for_transitions(F&& f, std::uint16_t c) const {
-        if(auto it = transitions_.find(c); it != transitions_.end()) {
-            f(it->second);
-        }
+    template<class F> void for_transitions(F&& f, std::uint16_t c) const {
+        if (auto it = transitions_.find(c); it != transitions_.end()) f(it->second);
     }
-    
+
     // F: void(std::uint16_t, const DFANode*)
-    template <class F>
-    void for_transitions(F&& f) const {
-        for (auto &[c, to] : transitions_) {
-                f(c, to);
-        }
+    template<class F> void for_transitions(F&& f) const {
+        for (auto& [c, to] : transitions_) f(c, to);
     }
 
     bool is_accepting() const { return accepting_; }