--- a/dialects/regex/pass/dfa2matcher.cpp
+++ b/dialects/regex/pass/dfa2matcher.cpp
@@ -3,6 +3,8 @@
 #include <algorithm>
 #include <sstream>
 
+#include <absl/container/flat_hash_map.h>
+
 #include "thorin/def.h"
 
 #include "dialects/core/core.h"
@@ -35,9 +37,9 @@ std::string state_to_name(const automaton::DFANode* state) {
 }
 
 // std::vector<std::pair<nat_t, nat_t>> merge_ranges(World& w, const std::vector<std::pair<nat_t, nat_t>>& old_ranges);
-std::unordered_map<const automaton::DFANode*, std::vector<std::pair<nat_t, nat_t>>>
+absl::flat_hash_map<const automaton::DFANode*, std::vector<std::pair<nat_t, nat_t>>>
 transitions_to_ranges(World& w, const automaton::DFANode* state) {
-    std::unordered_map<const automaton::DFANode*, std::vector<std::pair<nat_t, nat_t>>> state2ranges;
+    absl::flat_hash_map<const automaton::DFANode*, std::vector<std::pair<nat_t, nat_t>>> state2ranges;
     state->for_transitions([&](std::uint16_t transition, const automaton::DFANode* next_state) {
         if (!state2ranges.contains(next_state))
             state2ranges.emplace(next_state, std::vector<std::pair<nat_t, nat_t>>{
@@ -69,10 +71,10 @@ Ref match_range(Ref c, nat_t lo, nat_t hi) {
     return w.call(core::bit2::and_, w.lit_nat(2), w.tuple({below_hi, above_lo}));
 }
 
-std::unordered_map<const automaton::DFANode*, Ref>
+absl::flat_hash_map<const automaton::DFANode*, Ref>
 create_check_match_transitions_from(Ref c, const automaton::DFANode* state) {
     World& w = c->world();
-    std::unordered_map<const automaton::DFANode*, Ref> state2check;
+    absl::flat_hash_map<const automaton::DFANode*, Ref> state2check;
 
     auto state2ranges = transitions_to_ranges(w, state);
 
@@ -95,7 +97,7 @@ Ref dfa2matcher(World& w, const automaton::DFA& dfa, Ref n) {
     dfa.dump("dfa_to_match");
 #endif
     auto states = dfa.get_reachable_states();
-    std::unordered_map<const automaton::DFANode*, Lam*> state2matcher;
+    absl::flat_hash_map<const automaton::DFANode*, Lam*> state2matcher;
 
     // ((mem: %mem.M, string: Str n, pos: .Idx n), .Cn [%mem.M, .Bool, .Idx n])
     auto matcher_con  = w.cn(w.sigma({w.annex<mem::M>(), w.type_bool(), w.type_idx(n)}));