--- a/dialects/regex/pass/dfa2matcher.h
+++ b/dialects/regex/pass/dfa2matcher.h
@@ -1,0 +1,8 @@
+#pragma once
+#include "thorin/world.h"
+
+#include "dialects/regex/pass/dfa.h"
+
+namespace thorin::regex {
+Ref dfa2matcher(World& w, const automaton::DFA& dfa, Ref n);
+} // namespace thorin::regex