… dialects/CMakeLists.txt
… dialects/regex/pass/dfamin.h
… dialects/regex/pass/lower_regex.cpp
… dialects/regex/pass/nfa.cpp
… dialects/regex/pass/nfa.h
… dialects/regex/pass/nfa2dfa.cpp
… dialects/regex/pass/nfa2dfa.h
… dialects/regex/pass/regex2nfa.cpp
… dialects/regex/pass/regex2nfa.h
… dialects/regex/range_helper.cpp
… dialects/regex/range_helper.h
… dialects/regex/regex.thorin
… gtest/CMakeLists.txt
… gtest/automaton.cpp
… lit/regex/match_mail.thorin
… lit/regex/match_manual.thorin
… lit/regex/match_test10.thorin
… lit/regex/match_wplusa.thorin